+++ TGEngine/gamecontent/Material.cpp
@@ -11,7 +11,7 @@ namespace tge::gmc {
 	void loadMaterials(File file, ResourceDescriptor* desc, uint32_t size) {
 		for (uint32_t i = 0; i < size; i++) {
 			if (ftell(file) != desc[i].offset)
-				fseek(file, desc[i].offset, SEEK_SET);
+				fseek(file, (long)desc[i].offset, SEEK_SET);
 			fread(&createdMaterials[i], sizeof(char), desc[i].size, file);
 		}
 	}
